Método listarAtividades movido de DetalhesEventoActivity para AtividadeDAO
